split synthetic dataset
